requirements.txt
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==43.0.3
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
httplib2==0.22.0
idna==3.10
oauth2client==4.1.3
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
PyDrive2==1.21.3
pyOpenSSL==24.2.1
pyparsing==3.2.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
uritemplate==4.1.1
urllib3==2.3.0
uuid==1.30
